\name{airpass}
\docType{data}
\alias{airpass}
\title{International Airline Passengers}
\description{
Number of international airline passengers (in thousands) recorded monthly from January 1949 to December 1960.
}
\usage{data (airpass)}
\format{
A time series with 144 observations.
}
\keyword{datasets}






