% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor.plot.R
\name{autocorPlot}
\alias{autocor.plot}
\alias{autocorPlot}
\title{Autocorrelation Plot}
\usage{
autocorPlot(fit, main = "Current vs Lagged residuals", ...)
}
\arguments{
\item{fit}{output from the function 'lm()'.}

\item{main}{the plot title.}

\item{\ldots}{extra parameters to be passed to the \code{plot} function.}
}
\value{
Plots current vs lagged residuals along with quadrants dividing
these residuals about the value zero.
}
\description{
Plots current vs lagged residuals along with quadrants dividing these
residuals about the value zero.
}
\note{
\code{autocor.plot} is deprecated inline with our new policy of removing periods from function names.
}
\examples{

data(airpass.df)
time = 1:144
airpass.fit = lm(passengers ~ time, data = airpass.df)
autocorPlot(airpass.fit)

}
\keyword{hplot}

