% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{books.df}
\alias{books.df}
\title{Books Data}
\format{
The data frame consists of 400 observations on 2 variables.
\describe{
\item{length}{Integer sentence length, measured as the number of words in the sentence.}
\item{book}{Factor giving the book from which the sentence was sampled
  (4.Past.Mid, Dark.Half, Disclosure, Eye.Drag, Rising.Sun, Shining, Stand,
  T.Knock).}
}
}
\description{
This data consists of 50 sentence lengths from each of 8 books. The books
\dQuote{Disclosure} and \dQuote{Rising Sun} were written by Michael
Crichton, whilst the others \dQuote{Four Past Midnight}, \dQuote{The Dark
Half}, \dQuote{ Eye of the Dragon}, \dQuote{The Shining}, \dQuote{The Stand}
and \dQuote{The Tommy-Knockers} were written by Stephen King. The pages and
sentences were chosen using a multistage design where the pages were
selected at random, and then sentences within each page were selected at
random. These data were collected by James Curran.
}
\keyword{datasets}
