% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxqq.r
\name{boxqq}
\alias{boxqq}
\title{Deprecated box plots and normal quantile-quantile plots}
\usage{
boxqq(formula, ...)
}
\arguments{
\item{formula}{A symbolic specification of the form \code{x ~ g} can be given,
indicating the observations in the vector \code{x} are to be grouped according to
the levels of the factor \code{g}. \code{NA}'s are allowed in the data.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link{par}}).}
}
\value{
Returns the plot.
}
\description{
`boxqq()` is deprecated and is no longer exported. It draws boxplots
and normal quantile-quantile plots of `x` for each level of the grouping
variable `g`.
}
\note{
This is a legacy teaching helper retained for compatibility with
  older course material. New teaching material should prefer current
  diagnostic plotting workflows.
}
\keyword{hplot}
