% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{butterfat.df}
\alias{butterfat.df}
\title{Butterfat Data}
\format{
A data frame with 100 observations on 3 variables.
\describe{
\item{Butterfat}{numeric. Mean percentage of butterfat per cow}
\item{Breed}{factor. Breed (ayrshire, canadian, guernesy, holst.fres, jersey)}
\item{Age}{factor. Age group (2yo, mature)}
}
}
\source{
A Handbook of Small Data Sets
}
\description{
This data gives the mean percentage of butterfat produced by different
Canadian pure-bred diary cattle. There are five different breeds and two age
groups, two years old and greater than five years old. For each combination
of breed and age, there are measurements for 10 cows.
}
\references{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski,
E. (1994). \emph{A Handbook of Small Data Sets}. Boca Raton, Florida:
Chapman and Hall/CRC.

Sokal, R.R. and Rohlf, F.J. (1981). \emph{Biometry}, 2nd edition. San
Francisco: W.H. Freeman, 368.
}
\keyword{datasets}
