% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{camplake.df}
\alias{camplake.df}
\title{Age and Length of Camp Lake Bluegills}
\format{
A data frame with 66 observations on 3 variables.
\describe{
\item{Age}{Numeric age of the fish, in years.}
\item{Scale.Radius}{Numeric radius of the key scale, in hundredths of a millimetre.}
\item{Length}{Numeric length at capture, in millimetres.}
}
}
\description{
66 bluegills were captured from Camp Lake, Minnesota. For each bluegill we
have the length of the fish, its age in years and its age in scale radius.
}
\keyword{datasets}
