% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{chalk.df}
\alias{chalk.df}
\title{Chalk Data}
\format{
A data frame with 66 observations on 3 variables.
\describe{
\item{Density}{Numeric density of the chalk.}
\item{Lab}{Integer laboratory identifier.}
\item{Chalk}{Factor giving the chalk brand tested (A, B).}
}
}
\description{
These data involve 11 laboratories and 2 brands of chalk. The laboratories
tested the density of the chalk. The main interest was whether the different
laboratories yielded the same density for the two different types of chalk.
}
\keyword{datasets}
