% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{computer.df}
\alias{computer.df}
\title{Computer Questionnaire}
\format{
A data frame with 19 observations on 2 variables.
\describe{
\item{score}{numeric. Questionnaire score}
\item{selfassess}{ordered factor. Self-assessed level of expertise (1 = low, 2 = medium, 3 = high)}
}
}
\description{
Data from a test to see if a questionnaire was properly designed. The
questionnaire measures managers' technical knowledge of computers. The test
has 19 managers complete the questionnaire as well as rate their own
technical expertise.
}
\keyword{datasets}
