% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{course2way.df}
\alias{course2way.df}
\title{Exam Mark, Gender and Attendance for Stats 20x Summer School Students}
\format{
A data frame with 40 observations on 3 variables.
\describe{
\item{Exam}{numeric. Final exam mark (out of 100)}
\item{Gender}{factor. Gender (Female, Male)}
\item{Attend}{factor. Regularly attended or not (No, Yes)}
}
}
\description{
Data from a summer school Stats 20x course. Each observation represents a
single student. It is of interest to see if there is a relationship between
a student's final examination mark and both their gender and whether they
regularly attend lectures.
}
\keyword{datasets}
