\name{diamonds.df}
\docType{data}
\alias{diamonds.df}
\title{Prices and Weights of Diamonds}
\description{
Prices of ladies' diamond rings from a Singaporean retailer and the weight of their diamond stones.
}
\usage{data (diamonds.df)}
\format{
A data frame with 48 observations on 2 variables.
\tabular{rlll}{
[,1] \tab price \tab numeric \tab Price of ring (Singapore dollars) \cr
[,2] \tab weight  \tab numeric \tab Weight of Diamond (carats)
}
}
\keyword{datasets}

