% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateContrasts.R
\name{estimateContrasts}
\alias{estimateContrasts}
\title{Contrast Estimates}
\usage{
estimateContrasts(
  contrast.matrix,
  fit,
  row = TRUE,
  alpha = 0.05,
  L = NULL,
  FUN = identity
)
}
\arguments{
\item{contrast.matrix}{A matrix of contrast coefficients. Separate rows of
the matrix contain the contrast coefficients for that particular contrast,
and a column for each level of the factor.}

\item{fit}{Output from the [lm()] function.}

\item{row}{If `TRUE`, and the ANOVA is two-way, then contrasts in the row
effects are printed, otherwise contrasts in the column effects are printed.
Ignored if the ANOVA is one-way.}

\item{alpha}{The nominal error rate for the multiple confidence intervals.}

\item{L}{Number of contrasts. If `NULL`, `L` will be set to the number of rows
in the contrast matrix, otherwise `L` will be as specified.}

\item{FUN}{Optional function to be applied to estimates and confidence
intervals. Typically used for back-transformation operations.}
}
\value{
Returns a matrix whose rows correspond to the different contrasts
  being estimated and whose columns correspond to the point estimate of the
  contrast, the Tukey lower and upper limits of the confidence interval, the
  unadjusted p-value, and the Tukey and Bonferroni p-values.
}
\description{
Calculates and prints Tukey multiple confidence intervals for contrasts in
one or two-way ANOVA.
}
\note{
This function is no longer exported as it should never be called by the
  user. It will ultimately be removed.
}
\examples{
## computer data:
data(computer.df)
computer.df = within(computer.df, {selfassess = factor(selfassess)})
computer.fit = lm(score ~ selfassess, data = computer.df)
contrast.matrix = matrix(c(-1 / 2, -1 / 2, 1), byrow = TRUE, nrow = 1, ncol = 3)
contrast.matrix
s20x:::estimateContrasts(contrast.matrix, computer.fit)
}
\seealso{
[summary1way()], [summary2way()], [multipleComp()]
}
\keyword{models}
