% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslm.R
\name{extractTslmFit}
\alias{extractTslmFit}
\title{Extract the underlying tslm fit}
\usage{
extractTslmFit(model)
}
\arguments{
\item{model}{a model object.}
}
\value{
A fitted model object.
}
\description{
Return the underlying fitted model from a `tslm` object, or the input model unchanged.
}
\keyword{internal}
