\name{fire.df}
\docType{data}
\alias{fire.df}
\title{Fire Damage and Distance from the Fire Station}
\description{
House damage and distance from the fire station, of 15 house fires. Data collected by an insurance company for homes in a particular area.
}
\usage{data (fire.df)}
\format{
A data frame with 15 observations on 2 variables.
\tabular{rlll}{
[,1] \tab damage \tab numeric \tab Damage (\$000s)\cr
[,2] \tab distance \tab numeric \tab Distance from the fire station (miles)
}
}
\keyword{datasets}

