% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{fruitfly.df}
\alias{fruitfly.df}
\title{Fruitfly Data}
\format{
A data frame with 75 observations on 2 variables. \tabular{rlll}{
[,1] \tab fecundity \tab numeric \tab Number of eggs laid, per day, per
fruitfly \cr [,2] \tab strain \tab factor \tab Strain of fruitfly (NS, RS,
SS) }
}
\source{
A Handbook of Small Data Sets
}
\description{
This data gives fecundity for female fruitflies, Drosophila melanogaster.
The fecundity is the number of eggs laid, per day, for the fruitfly's first
14 days of life. There are three strains: A control group, NS, Nonselected
Strain, as well as RS, a strain bred for resistance to DDT and SS, a strain
bred for susceptibility to DDT. Each strain contains 25 measurements. It is
of interest to compare the level of fecundity across strains.
}
\references{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski,
E. (1994). \emph{A Handbook of Small Data Sets}. Boca Raton, Florida:
Chapman and Hall/CRC.

Sokal, R.R. and Rohlf, F.J. (1981). \emph{Biometry}, 2nd edition. San
Francisco: W.H. Freeman, 239.
}
\keyword{datasets}
