% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslm.R
\name{getTslmCoefficientTable}
\alias{getTslmCoefficientTable}
\title{Get a tslm coefficient table}
\usage{
getTslmCoefficientTable(fitSummary)
}
\arguments{
\item{fitSummary}{summary output from the underlying fitted model.}
}
\value{
A coefficient matrix.
}
\description{
Extract a coefficient table from either an `lm` or `gls` summary object.
}
\keyword{internal}
