% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslm.R
\name{getTslmResidualDf}
\alias{getTslmResidualDf}
\title{Get tslm residual degrees of freedom}
\usage{
getTslmResidualDf(fit, fitSummary)
}
\arguments{
\item{fit}{the underlying fitted model.}

\item{fitSummary}{summary output from the fitted model.}
}
\value{
The residual degrees of freedom, or `NA_integer_` if unavailable.
}
\description{
Extract residual degrees of freedom from a fitted model and its summary.
}
\keyword{internal}
