% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslm.R
\name{getTslmTimeValues}
\alias{getTslmTimeValues}
\title{Get tslm time values}
\usage{
getTslmTimeValues(object, nResiduals)
}
\arguments{
\item{object}{a fitted `tslm` object.}

\item{nResiduals}{expected number of residuals.}
}
\value{
A vector of time or observation-order values.
}
\description{
Extract or reconstruct the time values used by `tslm` diagnostic plots.
}
\keyword{internal}
