% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersion.R
\name{getVersion}
\alias{getVersion}
\title{s20x Package Version Number}
\usage{
getVersion()
}
\description{
Returns the version number of the s20x package. This is useful if a student
has problems running commands and the maintainer needs to check the version
number.
}
\examples{
getVersion()

}
\keyword{debugging}
