\name{house.df}
\docType{data}
\alias{house.df}
\title{Sale and Advertised Prices of Houses}
\description{
A random sample of 100 houses recently sold in Mt Eden, Auckland. For each house we have the advertised price and the actual sale price.
}
\usage{data (house.df)}
\format{
A data frame with 100 observations on 2 variables.
\tabular{rlll}{
[,1] \tab advertised.price \tab numeric \tab Advertised price (\$)\cr
[,2] \tab sell.price \tab numeric \tab Final sale price (\$)
}
}
\keyword{datasets}

