% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{incomes.df}
\alias{incomes.df}
\title{Mean Family Incomes}
\format{
A data frame with 152 observations on 1 variable.
\describe{
\item{incomes}{Numeric mean family income, in thousands of dollars.}
}
}
\description{
Random sample of 152 families giving their mean income (1000s of dollars). The sample
was taken by an advertising agency over their area of operations.
}
\keyword{datasets}
