% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{lakemary.df}
\alias{lakemary.df}
\title{Ages and Lengths of Lake Mary Bluegills}
\format{
A data frame with 78 observations on 2 variables.
\describe{
\item{Age}{numeric. Age of the fish (years)}
\item{Length}{numeric. Length at capture (mm)}
}
}
\description{
The ages and lengths of 78 bluegills captured from Lake Mary, Minnesota.
}
\keyword{datasets}
