\name{mazda.df}
\docType{data}
\alias{mazda.df}
\title{Year and Price of Mazda Cars}
\description{
Prices and ages of 124 Mazda cars collected from the Melbourne Age newspaper in 1991.
}
\usage{data (mazda.df)}
\format{
A data frame with 124 observations on 2 variables.
\tabular{rlll}{
[,1] \tab price \tab numeric \tab Price (Australian dollars)\cr
[,2] \tab year  \tab numeric \tab Year of manufacture
}
}
\keyword{datasets}

