% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{mening.df}
\alias{mening.df}
\title{Monthly Notifications of Meningococcal Disease}
\format{
A data frame with 144 observations on 3 variables.
\describe{
\item{Month}{Factor giving the month of notification.}
\item{Year}{Factor giving the year of notification.}
\item{mening}{Numeric number of notifications of meningococcal disease.}
}
}
\description{
This data shows the monthly number of notifications meningococcal disease in
New Zealand from January 1990 to December 2001.
}
\keyword{datasets}
