\name{mergers.df}
\docType{data}
\alias{mergers.df}
\title{Merger Days}
\description{
A random selection of 38 consummated mergers from the USA, 1982, giving the number of days between the date the merger was announced and the date the merger became effective.
}
\usage{data (mergers.df)}
\keyword{datasets}

