% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{mozart.df}
\alias{mozart.df}
\title{Length of Mozart's Movements}
\format{
A data frame with 88 observations on 3 variables.
\describe{
\item{Time}{numeric. Time of each movement (seconds)}
\item{Movement}{factor. Movement (M1, M2, M3, M4)}
\item{Period}{factor. Period that the symphony was written (early, late)}
}
}
\description{
Length of movements from 11 of Mozart's early symphonies and 11 of his late
symphonies.
}
\keyword{datasets}
