\name{panel.cor}
\alias{panel.cor}
\title{A Panel Function for Showing Correlations}
\usage{panel.cor (x, y, digits=2, prefix="", cex.cor)}
\arguments{
\item{x}{A vector of values}
\item{y}{A vector of values}
\item{digits}{Number of significant digits}
\item{prefix}{Optional characters placed in front of each coefficient}
\item{cex.cor}{Character size argument}
}
\description{This function is called by pairs20x to show correlation coefficients in panels. Text size is shown proportional to the strength of the correlation.}

\seealso{
\code{"pairs", "pairs20x", "panel.smooth", "panel.hist"}
}
\keyword{hplot}

