\name{panel.hist}
\alias{panel.hist}
\title{A Panel Function for Plotting Histograms}
\usage{panel.hist (x, \dots)}
\arguments{
\item{x}{A vector of observations}
\item{\dots}{Additional arguments, parsed to "rect" which draws the histogram bars.}
}
\description{This function is called by pairs20x to draw histograms in panels.}

\seealso{
\code{"pairs", "pairs20x", "panel.smooth", "panel.cor"}
}
\keyword{hplot}













