% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{peru.df}
\alias{peru.df}
\title{Peruvian Indians}
\format{
A data frame with 39 observations on 5 variables.
\describe{
\item{age}{numeric. Subject's age}
\item{years}{numeric. Number of years since migration}
\item{weight}{numeric. Subject's weight (kg)}
\item{height}{numeric. Subject's height (mm)}
\item{BP}{numeric. Subject's systolic blood pressure (mm Hg; standard clinical unit in New Zealand).}
}
}
\description{
A random sample of Peruvian Indians born in the Andes mountains, but who
have since migrated to lower altitudes. The sample was collected to assess
the long term effects of altitude on blood pressure.
}
\keyword{datasets}
