\name{peru}
\docType{data}
\alias{peru}
\title{Peruvian Indians}
\description{
A random sample of Peruvian Indians born in the Andes mountains, but who have since migrated to lower altitudes. The sample was collected to assess the long term effects of altitude on blood pressure.
}
\usage{data (peru)}
\format{
A data frame with 39 observations on 5 variables.
\tabular{rlll}{
[,1] \tab age \tab numeric \tab Subject's age \cr
[,2] \tab years  \tab numeric \tab Number of years since migration \cr
[,3] \tab weight \tab numeric \tab Subject's weight (kg) \cr
[,4] \tab height \tab numeric \tab Subject's height (mm) \cr
[,5] \tab BP \tab numeric \tab Subject's systolic blood pressure (mm Hg)
}
}
\keyword{datasets}








