% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-package.R
\docType{package}
\name{s0x}
\alias{s0x}
\alias{s0x-package}
\title{s20x: Functions for University of Auckland Course STATS 201/208 Data Analysis}
\description{
A set of functions used in teaching STATS 201/208 Data Analysis at the University of Auckland. The functions are designed to make parts of R more accessible to a large undergraduate population who are mostly not statistics majors.
}
\author{
Andrew Balemi, Dineika Chandra, James Curran, Brant Deppa, Mike Forster, Brendan McArdle, Chris Wild
}

