% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{seeds.df}
\alias{seeds.df}
\title{Seeds Data}
\format{
A data frame with 48 observations on 3 variables.
\describe{
\item{Light}{Factor indicating whether the seeds were exposed to light (N = No, Y = Yes).}
\item{Water}{Integer amount of water, with higher levels corresponding to more water (1, 2, 3, 4, 5, 6).}
\item{Count}{Integer number of seeds that germinated, out of 100.}
}
}
\description{
These data record the number of seeds (out of 100) that germinated when
given different amounts of water. The seeds were either exposed to light or
kept in the dark. Four identical boxes were used for each combination of
water and light
}
\keyword{datasets}
