% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{sheep.df}
\alias{sheep.df}
\title{Sheep Data}
\format{
A data frame with 100 observations on 3 variables.
\describe{
\item{Weight}{Integer Weight of sheep (kilograms, kg).}
\item{Copper}{Factor indicating whether copper supplementation was given (No, Yes).}
\item{Cobalt}{Factor indicating whether cobalt supplementation was given (No, Yes).}
}
}
\description{
Weight measurements for sheep under combinations of copper and cobalt
supplementation.
}
\keyword{datasets}
