% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.r
\name{skewness}
\alias{skewness}
\title{Skewness Statistic}
\usage{
skewness(x, ...)
}
\arguments{
\item{x}{Vector containing the data.}

\item{\dots}{Additional arguments passed to [mean()] and [sd()], such as
`na.rm = TRUE`.}
}
\value{
The value of the skewness statistic.
}
\description{
Calculates the skewness statistic of the data in `x`. Values close to zero
correspond to reasonably symmetric data. Positive values indicate right-skewed
data, whereas negative values indicate left-skewed data.
}
\examples{
data(mergers.df)
skewness(mergers.df$mergerdays)

}
\keyword{univar}
