\name{skulls.df}
\docType{data}
\alias{skulls.df}
\title{Skulls Data}
\description{
Male Egyptian skulls from five different epochs. Each skull has had four measurements taken of it, BH, Basibregmatic Height, BL, Basialveolar Length, MB, Maximum Breadth and NH, Nasal Height. It is of interest to investigate the change in shape over time. A gradual change, would indicate inbreeding of the populuations.
}
\usage{data (skulls.df)}
\format{
A data frame with 600 observations on 3 variables.
\tabular{rlll}{
[,1] \tab breadth \tab integer \tab  \cr
[,2] \tab type  \tab factor \tab levels (BH, BL, MB, NH) \cr
[,3] \tab Year  \tab integer \tab The epoch the skull is from
}
}
\source{
The data was taken from:

Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. \& Ostrowski, E. (1994). \textit {A Handbook of Small Data Sets}. Boca Raton, Florida: Chapman and Hall/CRC.

Original source:

Thomson, A. \& Randall-Maciver, R. (1905). \textit {Ancient Races of the Thebaid}. Oxford: Oxford University Press.
}
\keyword{datasets}








