% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{snapper.df}
\alias{snapper.df}
\title{Snapper Weight Data}
\format{
A data frame with 844 observations on 2 variables.
\describe{
\item{len}{Numeric fork length in centimetres. Fork length is measured from the tip of the snout to the fork of the tail.}
\item{wgt}{Numeric weight of the fish, in kilograms.}
}
}
\source{
Russell Millar, University of Auckland.
}
\description{
Weight and length measurements of 844 snapper
(\href{https://en.wikipedia.org/wiki/Australasian_snapper}{Pagrus auratus})
caught in the Hauraki Gulf, near Auckland, New Zealand.
}
\keyword{datasets}
