% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{soyabean.df}
\alias{soyabean.df}
\title{Soya Bean Yields}
\format{
A data frame with 32 observations on 3 variables.
\describe{
\item{yield}{numeric. Yield of each plant}
\item{cultivar}{factor. Cultivar used (cult1, cult2, cult3, cult4)}
\item{planttime}{factor. Month of planting (Novemb, Decemb)}
}
}
\source{
Littler, R. University of Waikato
}
\description{
Data from an experiment to examine the effects of different planting times
on the yield of soya beans, given four different cultivars.
}
\keyword{datasets}
