% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{teach.df}
\alias{teach.df}
\title{Comparison of Three Teaching Methods}
\format{
A data frame with 30 observations on 3 variables.
\describe{
\item{lang}{numeric. Language test score after instruction}
\item{IQ}{numeric. Student's IQ}
\item{method}{factor. Teaching method (1, 2, 3)}
}
}
\description{
Data from an experiment to assess the impact of three different teaching
methods on language ability. 30 students were randomly allocated into three
groups, one for each method. The students' IQ before instruction and a
language test score after instruction were recorded.
}
\keyword{datasets}
