% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{thyroid.df}
\alias{thyroid.df}
\title{Effect of a New Drug on Thyroid Weights}
\format{
A data frame with 16 observations on 3 variables.
\describe{
\item{thyroid}{numeric. Weight of thyroid gland after 7 days (mg)}
\item{body}{numeric. Animal body weight before experiment began (g)}
\item{group}{factor. Animal's group (1 = control, 2 = drug)}
}
}
\description{
Data from an experiment to asses the effect of a new drug on the weight of
the thyroid gland using 16 laboratory animals. The animals were randomly
assigned into either a control group, or a treatment group, and each animal
had its bodyweight recorded at the beginning of the experiment and its
thyroid weight measured at the end of the experiment.
}
\keyword{datasets}
