\name{toothpaste.df}
\docType{data}
\alias{toothpaste.df}
\title{Crest Toothpaste}
\description{
Two random samples of households, one of households who purchase Crest toothpaste and one of households who do not. For each household the age is recorded of the person responsible for purchasing the toothpaste.
}
\usage{data (toothpaste.df)}
\format{
A data frame with 20 observations on 2 variables.
\tabular{rlll}{
[,1] \tab purchasers \tab numeric \tab Age of the person in the household responsible for purchases of Crest \cr
[,2] \tab nonpurchasers \tab numeric \tab Age of the person in the household responsible for purchases of other brands of toothpaste \cr
}
}
\keyword{datasets}

