\name{version}
\alias{version}
\title{s20x pacakge version number}
\usage{
version()
}
\description{
Returns the version number of the s20x package. This is useful if a
student is has problems runnning commands and the maintainer needs to
check the version number.
}
\examples{
version()
}

\keyword{debugging}












