% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{airpass.df}
\alias{airpass.df}
\title{International Airline Passengers}
\format{
A data frame with 144 rows and 4 variables:
\describe{
  \item{passengers}{Monthly total number of international airline passengers (in thousands).}
  \item{t}{Integer time index from 1 to 144.}
  \item{month}{Month of observation as a factor with levels \code{Jan} to \code{Dec}.}
  \item{year}{Year of observation as a factor with levels \code{1949} to \code{1960}.}
}
}
\description{
Number of international airline passengers (in thousands) recorded monthly
from January 1949 to December 1960.
}
\keyword{datasets}
