% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{airpass.df}
\alias{airpass.df}
\title{International Airline Passengers}
\format{
A time series with 144 observations.
}
\description{
Number of international airline passengers (in thousands) recorded monthly
from January 1949 to December 1960.
}
\keyword{datasets}
