% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{arousal.df}
\alias{arousal.df}
\title{Changes in Pupil Size with Emotional Arousal}
\format{
A data frame with 160 observations on 3 variables.
\describe{
\item{arousal}{numeric. Change in the subject's pupil size}
\item{gender}{factor. Subject's gender (female, male)}
\item{picture}{factor. Picture shown to subject (infant, landscape, nude female, nude male)}
}
}
\description{
Data from an experiment to measure the effect of different images on
emotional arousal, by measuring changes in pupil diameter. The experiment
used 20 males and 20 females. Images included a nude man, nude woman,
infant, and a landscape.
}
\keyword{datasets}
