% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{body.df}
\alias{body.df}
\title{Body Image and Ethnicity}
\format{
A data frame with 246 observations on 8 variables.
\describe{
\item{ethnicity}{factor. Subject's ethnicity (Asian, Europn, Maori, Pacific)}
\item{married}{.. .}
\item{bodyim}{factor. Subject's rating of themself (slight.uw, right, slight.ow, mod.ow, very.ow)}
\item{sm.ever}{.. .}
\item{weight}{.. .}
\item{height}{.. .}
\item{age}{.. .}
\item{stressgp}{.. .}
}
}
\description{
Data collected to examine how women from various ethnic groups rate their
body image. All subjects were slightly underweight for their body size.
}
\keyword{datasets}
