\name{books.df}
\docType{data}
\alias{books.df}
\title{Books Data}
\description{
 This data consists of 50 sentence lengths from each of 8 books. The
 books \dQuote{Disclosure} and \dQuote{Rising Sun} were written by
 Michael Crichton, whilst the others \dQuote{Four Past Midnight},
 \dQuote{The Dark Half}, \dQuote{ Eye of the Dragon}, \dQuote{The
   Shining}, \dQuote{The Stand} and \dQuote{The Tommy-Knockers} where
 written by Stephen King. The pages and sentences where chosen using a
 multistage design where the pages where selected at random, and then
 sentences within each page were selected at random. These data were
 collected by James Curran.
}
\usage{data (books.df)}
\format{
 The data frame consists of 400 observations
 on 2 variables.
\tabular{rlll}{
[,1] \tab length \tab integer \tab  \cr
[,2] \tab book  \tab factor \tab levels (4.Past.Mid, Dark.Half, Disclosure, Eye.Drag, \cr
\tab \tab \tab Rising.Sun, Shining, Stand, T.Knock)
}
}
\keyword{datasets}

