\name{bursary.df}
\docType{data}
\alias{bursary.df}
\title{Bursary Results for Auckland Secondary Schools}
\description{
Data for the 2001 Bursary results for 75 secondary schools in the Auckland area. For each school the decile rating of the school is recorded along with the percentage of eligible students who gain a B Bursary or better.
}
\usage{data (bursary.df)}
\format{
A data frame with 75 observations on 2 variables.
\tabular{rlll}{
[,1] \tab decile \tab numeric \tab Decile rating of the school \cr
[,2] \tab pass.rate  \tab numeric \tab Percentage of eligible students who gained a "B" Busary or better
}
}
\keyword{datasets}

