% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{bursary.df}
\alias{bursary.df}
\title{Bursary Results for Auckland Secondary Schools}
\format{
A data frame with 75 observations on 2 variables.
\describe{
\item{decile}{numeric. Decile rating of the school}
\item{pass.rate}{numeric. Percentage of eligible students who gained a 'B' Busary or better}
}
}
\description{
Data for the 2001 Bursary results for 75 secondary schools in the Auckland
area. For each school the decile rating of the school is recorded along with
the percentage of eligible students who gain a B Bursary or better.
}
\keyword{datasets}
