% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{chalk.df}
\alias{chalk.df}
\title{Chalk Data}
\format{
A data frame with 66 observations on 3 variables.
\describe{
\item{Density}{numeric. Density of the chalk}
\item{Lab}{integer. Laboratory where testing done}
\item{Chalk}{factor. Chalk tested (A, B)}
}
}
\description{
These data involve 11 laboratories and 2 brands of chalk. The laboratories
tested the density of the chalk. The main interest was whether the different
laboratories yielded the same density for the two different types of chalk.
}
\keyword{datasets}
