\name{chalk.df}
\docType{data}
\alias{chalk.df}
\title{Chalk Data}
\description{
These data involve 11 laboratories and 2 brands of chalk. The laboratories
tested the density of the chalk. The main interest was whether the different
laboratories yielded the same density for the two different types of chalk.
}
\usage{data (chalk.df)}
\format{
A data frame with 66 observations on 3 variables.
\tabular{rlll}{
[,1] \tab Density \tab numeric \tab Density of the chalk \cr
[,2] \tab Lab \tab integer \tab Laboratory where testing done \cr
[,3] \tab Chalk \tab factor \tab Chalk tested (A, B)
}
}
\keyword{datasets}

