\name{course2way.df}
\docType{data}
\alias{course2way.df}
\title{Exam Mark, Gender and Attendance for Stats 20x Summer School Students}
\description{
Data from a summer school Stats 20x course. Each observation represents a single student. It is of interest to see if there is a relationship between a student's final examination mark and both their gender and whether they regularly attend lectures.
}
\usage{data (course2way.df)}
\format{
A data frame with 40 observations on 3 variables.
\tabular{rlll}{
[,1] \tab Exam \tab numeric \tab Final exam mark (out of 100) \cr
[,2] \tab Gender \tab factor \tab Gender (Female, Male) \cr
[,2] \tab Attend \tab factor \tab Regularly attended or not (No, Yes)
}
}
\keyword{datasets}

