% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{diamonds.df}
\alias{diamonds.df}
\title{Prices and Weights of Diamonds}
\format{
A data frame with 48 observations on 2 variables.
\describe{
\item{price}{numeric. Price of ring (Singapore dollars)}
\item{weight}{numeric. Weight of Diamond (carats)}
}
}
\description{
Prices of ladies' diamond rings from a Singaporean retailer and the weight
of their diamond stones.
}
\keyword{datasets}
