% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{fire.df}
\alias{fire.df}
\title{Fire Damage and Distance from the Fire Station}
\format{
A data frame with 15 observations on 2 variables.
\describe{
\item{damage}{numeric. Damage (1000s of dollars)}
\item{distance}{numeric. Distance from the fire station (miles)}
}
}
\description{
House damage and distance from the fire station, of 15 house fires. Data
collected by an insurance company for homes in a particular area.
}
\keyword{datasets}
