% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{house.df}
\alias{house.df}
\title{Sale and Advertised Prices of Houses}
\format{
A data frame with 100 observations on 2 variables.
\describe{
\item{advertised.price}{numeric. Advertised price (dollars)}
\item{sell.price}{numeric. Final sale price (dollars)}
}
}
\description{
A random sample of 100 houses recently sold in Mt Eden, Auckland. For each
house we have the advertised price and the actual sale price.
}
\keyword{datasets}
