% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{house.df}
\alias{house.df}
\title{Sale and Advertised Prices of Houses}
\format{
A data frame with 100 observations on 2 variables. \tabular{rlll}{
[,1] \tab advertised.price \tab numeric \tab Advertised price (dollars)\cr [,2]
\tab sell.price \tab numeric \tab Final sale price (dollars) }
}
\description{
A random sample of 100 houses recently sold in Mt Eden, Auckland. For each
house we have the advertised price and the actual sale price.
}
\keyword{datasets}
