% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{incomes.df}
\alias{incomes.df}
\title{Mean Family Incomes}
\description{
Random sample of 152 families giving their mean income (1000s of dollars). The sample
was taken by an advertising agency over their area of operations.
}
\keyword{datasets}
