\name{lakemary.df}
\docType{data}
\alias{lakemary.df}
\title{Ages and Lengths of Lake Mary Bluegills}
\description{
The ages and lengths of 78 bluegills captured from Lake Mary, Minnesota.
}
\usage{data (lakemary.df)}
\format{
A data frame with 78 observations on 2 variables.
\tabular{rlll}{
[,1] \tab Age \tab numeric \tab Age of the fish (years)\cr
[,2] \tab Length  \tab numeric \tab Length at capture (mm)
}
}
\keyword{datasets}

