% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-package.R
\docType{data}
\name{larain.df}
\alias{larain.df}
\title{Los Angeles Rainfall}
\format{A time series with 66 observations.}
\description{
Annual rainfall (in inches) for Los Angeles from 1908 to 1973.
}
\keyword{datasets}
