% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout20x.R
\name{layout20x}
\alias{layout20x}
\title{Layout}
\usage{
layout20x(numRows = 1, numCols = 1)
}
\arguments{
\item{numRows}{number of rows in plot array}

\item{numCols}{number of columns in plot array}
}
\value{
Function returns no value
}
\description{
Allows an \code{numRows} by \code{numCols} matrix of plots to be displayed
in a single plot. If the function is called with no arguments, then the
plotting device layout will be reset to a single plot.
}
\note{
This function is deprecated. It will be removed in future versions of the package.
}
\examples{

data(course.df)
layout20x(1,2)
stripchart(course.df$Exam)
boxplot(course.df$Exam)

}
\keyword{device}
