% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{mazda.df}
\alias{mazda.df}
\title{Year and Price of Mazda Cars}
\format{
A data frame with 124 observations on 2 variables.
\describe{
\item{price}{numeric. Price (Australian dollars)}
\item{year}{numeric. Year of manufacture}
}
}
\description{
Prices and ages of 124 Mazda cars collected from the Melbourne Age newspaper
in 1991.
}
\keyword{datasets}
