\name{mening.df}
\docType{data}
\alias{mening.df}
\title{Monthly Notifications of Meningococcal Disease}
\description{
This data shows the monthly number of notifications meningococcal disease in New Zealand from January 1990 to December 2001.
}
\usage{data (mening.df)}
\format{
A data frame with 144 observations on 3 variables: Month, Year and mening.
}
\keyword{datasets}

