% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{mergers.df}
\alias{mergers.df}
\title{Merger Days}
\description{
A random selection of 38 consummated mergers from the USA, 1982, giving the
number of days between the date the merger was announced and the date the
merger became effective.
}
\keyword{datasets}
