% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modcheck.R
\name{modcheck}
\alias{modcheck}
\alias{modcheck.lm}
\title{Model checking plots}
\usage{
modcheck(x, ...)

\method{modcheck}{lm}(x, plotOrder = 1:4, args = list(predResArgs,
  normcheckArgs, cooksArgs), parVals = list(mfrow = c(2, 2), xaxs = "r", yaxs
  = "r", pty = "s", mai = c(0.2, 0.2, 0, 0.05)), ...)
}
\arguments{
\item{x}{a vector of observations, or the residuals from fitting a linear model.  Alternatively, a fitted \code{lm} object.
If \code{x} is a single vector, then the implicit assumption is that the mean (or null) model is being 
fitted, i.e. \code{lm(x ~ 1)} and that the data are best summarised by the sample mean.}

\item{plotOrder}{the order of the plots. 1: pred-res plot, 2: normal Q-Q plot, 3: histogram, 4: Cooks's
Distance plot.}
}
\description{
Plots four model checking plots: an pred-res plot (residuals against predicted values), a Normal
Quantile-Quantile (Q-Q) plot, a histogram of the residuals with a normal distribution super-imposed
and a Cook's Distance plot.
}
\section{Methods (by class)}{
\itemize{
\item \code{lm}: Model checking plots
}}

\examples{

# An exponential growth curve
e = rnorm(100, 0, 0.1)
x = rnorm(100)
y = exp(5 + 3 * x + e)
fit = lm(y ~ x)
modcheck(fit)

# An exponential growth curve with the correct transformation
fit = lm(log(y) ~ x)
modcheck(fit)

# Peruvian Indians data
data(peru.df)
modcheck(lm(BP ~ weight, data = peru.df))

}
\keyword{hplot}
