% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcheck.R
\name{modelcheck}
\alias{modelcheck}
\alias{modelcheck.lm}
\title{Model checking plots
Compact layout for model checking plots.}
\usage{
modelcheck(x, ...)

\method{modelcheck}{lm}(x, which = 1:3, mar = c(3, 4, 1.5, 4), ...)
}
\arguments{
\item{x}{The fitted model.}

\item{which}{The plot(s) to be drawn. Residuals versus fitted values
(\code{which = 1}), histogram and Q-Q plot of residuals
(\code{which = 2}), and Cook's distance plot (\code{which = 3}).}

\item{mar}{Margins applied to each selected plot.}

\item{\ldots}{any other arguments to pass to \code{\link{plot}}}
}
\value{
Draws diagnostic plots for teaching model checking. The function is
called for its plotting side effects and does not provide a stable data return
object.
}
\description{
Model checking plots
Compact layout for model checking plots.
}
\section{Methods (by class)}{
\itemize{
\item \code{modelcheck(lm)}: Model checking plots

}}
\examples{
data(peru.df)
lmFit = lm(BP ~ weight, data = peru.df)

# Plot residuals versus fitted values only
modelcheck(lmFit, 1)

# Plot residuals versus fitted values, histogram, and Q-Q plot
modelcheck(lmFit, 1:2)

# Plot all diagnostics
modelcheck(lmFit)
}
