% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-package.R
\docType{data}
\name{mozart.df}
\alias{mozart.df}
\title{Length of Mozart's Movements}
\format{
A data frame with 88 observations on 3 variables. \tabular{rlll}{
[,1] \tab Time \tab numeric \tab Time of each movement (seconds) \cr [,2]
\tab Movement \tab factor \tab Movement (M1, M2, M3, M4) \cr [,3] \tab
Period \tab factor \tab Period that the symphony was written (early, late) }
}
\description{
Length of movements from 11 of Mozart's early symphonies and 11 of his late
symphonies.
}
\keyword{datasets}
