\name{nail.df}
\docType{data}
\alias{nail.df}
\title{Nail Polish Data}
\description{
These data were collected to determine whether quick drying nail polish or
regular nail polish dried faster. The time for each type of nail polish to
dry was recorded.
}
\usage{data (nail.df)}
\format{
A data frame with 60 observations on 2 variables.
\tabular{rlll}{
[,1] \tab polish \tab factor \tab Type of polish (Regular, Quick) \cr
[,2] \tab dry \tab integer \tab Time (in seconds) for the polish to dry
}
}
\keyword{datasets}

