% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{nail.df}
\alias{nail.df}
\title{Nail Polish Data}
\format{
A data frame with 60 observations on 2 variables. \tabular{rlll}{
[,1] \tab polish \tab factor \tab Type of polish (Regular, Quick) \cr [,2]
\tab dry \tab integer \tab Time (in seconds) for the polish to dry }
}
\description{
These data were collected to determine whether quick drying nail polish or
regular nail polish dried faster. The time for each type of nail polish to
dry was recorded.
}
\keyword{datasets}
