% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{nzalc.df}
\alias{nzalc.df}
\title{Quarterly Alcohol Available for Consumption in New Zealand}
\format{
A data frame with quarterly observations on 4 variables.
\describe{
\item{year}{Integer Year.}
\item{month}{Factor Month at the end of the quarter.}
\item{volume}{Numeric volume available for consumption, in million litres.}
\item{category}{Factor Beverage category.}
}
}
\source{
Stats NZ, Alcohol available for consumption: Year ended December 2021.
}
\description{
Quarterly alcohol available for consumption in New Zealand from 1935 to 2021.
The data give volumes of alcoholic beverages available for consumption,
grouped into broad beverage categories.
}
\keyword{datasets}
