% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{oysters.df}
\alias{oysters.df}
\title{Oyster Abundances over Different Sites}
\format{
A data frame with 87 observations on 2 variables. \tabular{rlll}{
[,1] \tab Oysters \tab numeric \tab Number of oysters on each experimental
panel \cr [,2] \tab Site \tab factor \tab Location of the experimental
panels (GR = Georges River, PS1 = First Port Stephens Site, PS2 = Second
Port Stephens Site) }
}
\description{
Data from an experiment to determine the abundance of oysters recruiting
from three sites in two different estuaries in New South Whales. One in
Georges River and two in Port Stephens. The number of oysters were recorded
for 10 cm by 10 cm panels over a two year period.
}
\keyword{datasets}
