% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propslsd.new.R
\name{propslsd.new}
\alias{propslsd.new}
\title{LSD-Display Intervals}
\usage{
propslsd.new(crosstablist, conf.level = 0.95, arrowlength = 0.1, ...)
}
\arguments{
\item{crosstablist}{A list produced by 'crosstabs' or a matrix containing a
2-way table of counts (without marginal totals).}

\item{conf.level}{Confidence level of the intervals.}

\item{arrowlength}{Length of the arrows.}

\item{\dots}{Additional arguments (ignored).}
}
\description{
This function is called by rowdistr.
}
\seealso{
\code{'crosstabs', 'rowdistr'}
}
\keyword{htest}

