\name{rain.df}
\docType{data}
\alias{rain.df}
\title{Cloud Seeding and Levels of Rainfall}
\description{
Data from an experiment to see if seeding clouds with Silver Nitrate effects the amount of rainfall.
}
\usage{data (rain.df)}
\format{
A data frame with 50 observations on 2 variables.
\tabular{rlll}{
[,1] \tab rain \tab numeric \tab Amount of rain \cr
[,2] \tab seed  \tab factor \tab Whether the clouds are seeded or not (seeded, unseeded)
}
}
\source{
Chambers, Cleveland, Kleiner, Tukey. (1983). Graphical Methods for Data Analysis.
}
\keyword{datasets}

