% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{rain.df}
\alias{rain.df}
\title{Cloud Seeding and Levels of Rainfall}
\format{
A data frame with 50 observations on 3 variables.
\describe{
\item{rain}{Amount of rain, measured in acre-feet (the volume of water required to cover one acre of land to a depth of one foot).}
\item{rain_ML}{Amount of rain expressed in megalitres (ML); \code{rain_ML = rain * 1.23348184}.}
\item{seed}{factor. Whether the clouds are seeded or not (seeded, unseeded)}
}
}
\description{
Data from an experiment to see if seeding clouds with Silver Nitrate effects
the amount of rainfall.
}
\keyword{datasets}
