% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-package.R
\docType{package}
\name{s20x-package}
\alias{s20x}
\alias{s20x-package}
\title{s20x: Functions for University of Auckland Course STATS 201/208 Data Analysis}
\description{
A set of functions used in teaching STATS 201/208 Data Analysis at the University of Auckland. The functions are designed to make parts of R more accessible to a large undergraduate population who are mostly not statistics majors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STATS-UOA/s20x}
  \item Report bugs at \url{https://github.com/STATS-UOA/s20x/issues}
}

}
\author{
\strong{Maintainer}: James Curran \email{j.curran@auckland.ac.nz} (Wrote the original R package. Current maintainer.)

Authors:
\itemize{
  \item Brant Deppa (Wrote the original R scripts this package is derived from)
}

Other contributors:
\itemize{
  \item Rachel Fewster \email{r.fewster@auckland.ac.nz} [contributor]
  \item Russell Millar \email{r.millar@auckland.ac.nz} [contributor]
  \item Ben Stevenson \email{ben.stevenson@auckland.ac.nz} [contributor]
  \item Andrew Balemi \email{a.balemi@auckland.ac.nz} [contributor]
  \item Chris Wild \email{c.wild@auckland.ac.nz} [contributor]
  \item Sophie Jones [contributor]
  \item Dineika Chandra [contractor]
  \item Brendan McArdle [contractor]
}

}
\keyword{internal}
