\name{sheep.df}
\docType{data}
\alias{sheep.df}
\title{Sheep Data}
\description{Sheep Data}
\usage{data (sheep.df)}
\format{
A data frame with 100 observations on 3 variables.
\tabular{rlll}{
[,1] \tab Weight \tab integer \tab . \cr
[,2] \tab Copper \tab factor \tab levels (No, Yes) \cr
[,3] \tab Cobalt \tab factor \tab levels (No, Yes)
}
}
\keyword{datasets}

