% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{sheep.df}
\alias{sheep.df}
\title{Sheep Data}
\format{
A data frame with 100 observations on 3 variables.
\describe{
\item{Weight}{integer. Weight of sheep (kilograms, kg).}
\item{Copper}{factor. levels (No, Yes)}
\item{Cobalt}{factor. levels (No, Yes)}
}
}
\description{
Sheep Data
}
\keyword{datasets}
