% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-package.R
\docType{data}
\name{sheep.df}
\alias{sheep.df}
\title{Sheep Data}
\format{A data frame with 100 observations on 3 variables. \tabular{rlll}{
[,1] \tab Weight \tab integer \tab . \cr [,2] \tab Copper \tab factor \tab
levels (No, Yes) \cr [,3] \tab Cobalt \tab factor \tab levels (No, Yes) }}
\description{
Sheep Data
}
\keyword{datasets}
