% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.r
\name{skewness}
\alias{skewness}
\title{Skewness Statistic}
\usage{
skewness(x, ...)
}
\arguments{
\item{x}{vector containing the data.}

\item{\dots}{any other variables to be passed to \code{mean} and \code{sd}, e.g. 
\code{na.rm = TRUE}.}
}
\value{
Returns the value of the skewness.
}
\description{
Calculates the skewness statistic of the data in 'x'. Values close to zero
correspond to reasonably symmetric data, positive values of this measure
indicate right-skewed data whereas negative values indicate left-skewness.
}
\examples{

##Merger data:
data(mergers.df)
skewness(mergers.df$mergerdays)

}
\keyword{univar}
