% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{skulls.df}
\alias{skulls.df}
\title{Skulls Data}
\format{
A data frame with 150 observations on 2 variables.
\describe{
\item{measurement}{integer}
\item{year}{integer}
}
}
\source{
A Handbook of Small Data Sets
}
\description{
Male Egyptian skulls from five different epochs. Each skull has had four
measurements taken of it, BH, Basibregmatic Height, BL, Basialveolar Length,
MB, Maximum Breadth and NH, Nasal Height. It is of interest to investigate
the change in shape over time. A gradual change, would indicate inbreeding
of the populations. This data only includes the maximum breadth
measurements.
}
\references{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski,
E. (1994). \emph{A Handbook of Small Data Sets}. Boca Raton, Florida:
Chapman and Hall/CRC.

Thomson, A. and Randall-Maciver, R. (1905). \emph{Ancient Races of the
Thebaid}. Oxford: Oxford University Press.
}
\keyword{datasets}
