% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripqq.r
\name{stripqq}
\alias{stripqq}
\alias{stripqq.formula}
\title{Deprecated strip charts and normal quantile-quantile plots}
\usage{
stripqq(formula, ...)

\method{stripqq}{formula}(formula, data = NULL, ...)
}
\arguments{
\item{formula}{A symbolic specification of the form \code{x ~ g} can be given,
indicating the observations in the vector `x' are to be grouped according to
the levels of the factor `g'. \code{NA}s are allowed in the data.}

\item{data}{An optional data frame in which to evaluate the formula}

\item{\ldots}{Optional arguments that are passed to the \code{stripchart} function.}
}
\description{
`stripqq()` is deprecated and is no longer exported. It draws strip charts and normal quantile quantile plots of x for each value of
the grouping variable g
}
\section{Methods (by class)}{
\itemize{
\item \code{stripqq(formula)}: Strip charts and normal quantile-quantile plots

}}
\note{
This is a legacy teaching helper retained for compatibility with
  older course material. New teaching material should prefer current
  diagnostic plotting workflows.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# This legacy helper is no longer exported.
\dontshow{\}) # examplesIf}
}
\keyword{hplot}
