% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripqq.r
\name{stripqq}
\alias{stripqq}
\alias{stripqq.formula}
\title{Strip charts and normal quantile-quantile plots}
\usage{
stripqq(formula, ...)

\method{stripqq}{formula}(formula, data = NULL, ...)
}
\arguments{
\item{formula}{A symbolic specification of the form \code{x ~ g} can be given,
indicating the observations in the vector `x' are to be grouped according to
the levels of the factor `g'. \code{NA}s are allowed in the data.}

\item{data}{An optional data frame in which to evaluate the formula}

\item{\ldots}{Optional arguments that are passed to the \code{stripchart} function.}
}
\description{
Draws strip charts and normal quantile quantile plots of x for each value of
the grouping variable g
}
\section{Methods (by class)}{
\itemize{
\item \code{stripqq(formula)}: Strip charts and normal quantile-quantile plots

}}
\note{
This function is deprecated and will be removed in later versions of the pacakge.
}
\examples{

## Zoo data
data(zoo.df)
stripqq(attendance~day.type, data = zoo.df)

}
\keyword{hplot}
