% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{technitron.df}
\alias{technitron.df}
\title{Technitron Salary Information}
\format{
A data frame with 46 observations on 8 variables.
\describe{
\item{salary}{numeric. Annual Salary (dollars)}
\item{yrs.empl}{numeric. Number of years employed at Technitron}
\item{prior.yrs}{numeric. Number of years prior experience}
\item{educ}{numeric. Years of education after high school}
\item{id}{numeric. Company identification number}
\item{gender}{numeric. Gender (0 = female, 1 = male)}
\item{dept}{numeric. Department employee works in (1 = Sales, 2 = Purchasing, 3 = Advertising, 4 = Engineering)}
\item{super}{numeric. Number of employees supervised}
}
}
\description{
Salary information for all salaried employees of the Technitron Company.
}
\keyword{datasets}
