\name{thyroid.df}
\docType{data}
\alias{thyroid.df}
\title{Effect of a New Drug on Thyroid Weights}
\description{
Data from an experiment to asses the effect of a new drug on the weight of the thyroid gland using 16 laboratory animals. The animals were randomly assigned into either a control group, or a treatment group, and each animal had its bodyweight recorded at the beginning of the experiment and its thyroid weight measured at the end of the experiment.
}
\usage{data(thyroid.df)}
\format{
A data frame with 16 observations on 3 variables.
\tabular{rlll}{
[,1] \tab thyroid \tab numeric \tab Weight of thyroid gland after 7 days (mg)\cr
[,2] \tab body  \tab numeric \tab Animal body weight before experiment began (g) \cr
[,3] \tab group  \tab factor \tab Animal's group (1 = control, 2 = drug)
}
}
\keyword{datasets}

