% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{toothpaste.df}
\alias{toothpaste.df}
\title{Crest Toothpaste}
\format{
A data frame with 20 observations on 2 variables.
\describe{
\item{purchasers}{numeric. Age of the person in the household responsible for purchases of Crest}
\item{nonpurchasers}{numeric. Age of the person in the household responsible for purchases of other brands of toothpaste}
}
}
\description{
Two random samples of households, one of households who purchase Crest
toothpaste and one of households who do not. For each household the age is
recorded of the person responsible for purchasing the toothpaste.
}
\keyword{datasets}
