% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s20x-data.R
\docType{data}
\name{zoo.df}
\alias{zoo.df}
\title{Zoo Attendance during an Advertising Campaign}
\format{
A data frame with 440 observations on 6 variables.
\describe{
\item{attendance}{numeric. Number of visitors}
\item{time}{numeric. Time in days since the start of the study}
\item{sun.yesterday}{numeric. Hours of sunshine the previous day}
\item{tv.adds}{numeric. Average spending on TV advertising in the previous week (1000s of dollars per day)}
\item{nice.day}{factor. Assessment based on number of hours of sunshine (0 = No, 1 = Yes)}
\item{day.type}{factor. Type of day (1 = ordinary weekday, 2 = weekend day, 3 = school holiday weekday, 4 = public holday)}
}
}
\description{
Data for 455 days of attendance records for Auckland Zoo, from January 1,
1993. Note that only 440 values are given due to missing values. It was of
interest to assess whether an advertising campaign was effective in
increasing attendance.
}
\keyword{datasets}
